requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
blessings==1.7
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
colorful==0.5.4
configparser==5.0.2
cycler==0.10.0
docker-pycreds==0.4.0
filelock==3.0.12
flatbuffers==1.12
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.17
google-api-core==1.28.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
gpustat==0.6.0
grpcio==1.34.1
h5py==3.1.0
hiredis==2.0.0
idna==2.10
importlib-metadata==4.0.1
jsonschema==3.2.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
msgpack==1.0.2
multidict==5.1.0
numpy==1.19.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencensus==0.7.13
opencensus-context==0.1.2
opencv-python==4.5.2.52
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pathtools==0.1.2
Pillow==8.2.0
progressbar2==3.53.1
prometheus-client==0.10.1
promise==2.3
protobuf==3.17.0
psutil==5.8.0
py-spy==0.3.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-utils==2.5.6
pytz==2021.1
PyYAML==5.4.1
ray==1.3.0
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sentry-sdk==1.1.0
shortuuid==1.0.1
six==1.15.0
smmap==4.0.0
subprocess32==3.5.4
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.4
wandb==0.10.30
Werkzeug==2.0.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.1
